+++ custom_model_runner/CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-#### [1.4.6] - 2020-12-04
+#### [1.4.6] - in progress
+##### Added
+- **/predictions** and **/predictionsUnstructured** endpoints as aliases for **/predict** and **/predictUnstructured**
+- handling the case when input sent as binary data
 ##### Fixed
 - Validation of numeric multiclass class labels should always compare as strings